Lazy annotation applied to basket Item service impl injection in order to avoid circular reference. I will implement a helper class to handle methods that are common for both basket service and basket item service

+++ src/main/java/com/fabiolima/online_shop/service_implementation/BasketServiceImpl.java
@@ -1,29 +1,44 @@
-package com.fabiolima.online_shop.service;
+package com.fabiolima.online_shop.service_implementation;
 
 import com.fabiolima.online_shop.exceptions.ForbiddenException;
 import com.fabiolima.online_shop.exceptions.NotFoundException;
 import com.fabiolima.online_shop.model.Basket;
+import com.fabiolima.online_shop.model.BasketItem;
 import com.fabiolima.online_shop.model.User;
 import com.fabiolima.online_shop.model.enums.BasketStatus;
 import com.fabiolima.online_shop.repository.BasketRepository;
+import com.fabiolima.online_shop.service.BasketItemService;
+import com.fabiolima.online_shop.service.BasketService;
+import com.fabiolima.online_shop.service.UserService;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
+/**
+ * clearBasket must be used in case the user wants to keep the basket open, but want to delete all items in it.
+ * deleteBasket must be used in case the user wants to delete the basket
+ */
 @Service
 public class BasketServiceImpl implements BasketService {
 
     private final BasketRepository basketRepository;
     private final UserService userService;
+    private final BasketItemService basketItemService;
 
     @Autowired
     public BasketServiceImpl (BasketRepository basketRepository,
-                              UserService userService){
+                              UserService userService,
+                              @Lazy BasketItemService basketItemService){
         this.basketRepository = basketRepository;
         this.userService = userService;
+       this.basketItemService = basketItemService;
     }
+
 
     @Override
     @Transactional
@@ -85,10 +100,19 @@ public class BasketServiceImpl implements BasketService {
 
     @Override
     public Basket deleteBasketById(Long userId, Long basketId) {
-
+        // validate basket
         Basket reference = validateAndFetchBasket(userId, basketId);
+
+        //check if its status is checked_out
         if(reference.getBasketStatus() == BasketStatus.CHECKED_OUT)
             throw new ForbiddenException("Cannot delete a checked out basket.");
+
+        //if not, clear the basket, giving back to stock all the quantity in items
+        //clearBasket(basketId);
+
+        //delete basket if it is empty
+        if (!reference.getBasketItems().isEmpty())
+            throw new ForbiddenException("Basket must be empty before deleting it.");
         basketRepository.deleteById(basketId);
         return reference;
     }
@@ -99,11 +123,36 @@ public class BasketServiceImpl implements BasketService {
                 .orElseThrow(() -> new NotFoundException(String.format("Basket with Id %d not found",basketId)));
     }
 
-    @Override
+    @Override//I have to pass all the quantity items back to product
+    @Transactional
     public Basket clearBasket(Long basketId) {
+        //find the basket
         Basket theBasket = findBasketById(basketId);
-        theBasket.getBasketItems().clear();
-        return basketRepository.save(theBasket);
+
+        //get the list of items and pass item by item to removeItemFromBasket method
+
+        List<BasketItem> listOfItem = theBasket.getBasketItems();
+        for(BasketItem item : listOfItem){
+            Long itemId = item.getId();
+            //this method will remove item from basket, update stock in database, update basket in database and
+            //delete items from database
+            //basketItemService.removeItemFromBasket(basketId,itemId);
+        }
+        return theBasket;
+    }
+
+    @Override
+    @Scheduled(fixedRate = 60000) // run every 60 seconds
+    public void deleteExpiredBasket() {
+
+        LocalDateTime expirationTime = LocalDateTime.now().minusMinutes(15);
+
+        List<Basket> expiredBaskets = basketRepository.findAllByLastUpdatedBefore(expirationTime);
+
+        if (!expiredBaskets.isEmpty()) {
+            basketRepository.deleteAll(expiredBaskets);
+            System.out.println("Deleted expired baskets: " + expiredBaskets.size());
+        }
     }
 
     @Override
@@ -121,7 +170,7 @@ public class BasketServiceImpl implements BasketService {
                 .sum();
     }
 
-    protected Basket validateAndFetchBasket(Long userId, Long basketId){
+    public Basket validateAndFetchBasket(Long userId, Long basketId){
 
         return basketRepository.findBasketByIdAndUserId(basketId, userId)
                 .orElseThrow(() -> new NotFoundException(
